[relayer]: extract outbound_consensus module from pallet lib.rs
